+++ examples/pragma.js
@@ -1,3 +1,0 @@
-var pragma = {
-  foo: 1
-};