+++ examples/unknown_pragma.js
@@ -1,0 +1,1 @@
+var unknown_pragma = {};